Uc12 - adiciona a opção de download de termo de estágio para o coordenador

+++ src/main/java/br/edu/ifpb/pweb2/estagion/repositories/InternshipRepository.java
@@ -3,5 +3,5 @@ package br.edu.ifpb.pweb2.estagion.repositories;
 import br.edu.ifpb.pweb2.estagion.model.Internship;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface InternshipRepository extends JpaRepository<Internship, Long> {
+public interface InternshipRepository extends JpaRepository<Internship, Integer> {
 }
